format retransform-class-scenario
