+++ test/plugin/pom.xml
@@ -184,7 +184,7 @@
                         <sourceDirectory>scenarios/redisson-scenario</sourceDirectory>
                         <sourceDirectory>scenarios/resttemplate-4.x-scenario</sourceDirectory>
                         <sourceDirectory>scenarios/retransform-class-scenario</sourceDirectory>
-<!--                        <sourceDirectory>scenarios/retransform-class-tomcat-scenario</sourceDirectory>-->
+                        <sourceDirectory>scenarios/retransform-class-tomcat-scenario</sourceDirectory>
                         <sourceDirectory>scenarios/servicecomb-0.x-scenario</sourceDirectory>
                         <sourceDirectory>scenarios/servicecomb-1.x-scenario</sourceDirectory>
                         <sourceDirectory>scenarios/shardingsphere-3.x-scenario</sourceDirectory>
@@ -194,7 +194,7 @@
                         <sourceDirectory>scenarios/shardingsphere-4.x-scenario</sourceDirectory>
                         <sourceDirectory>scenarios/sofarpc-scenario</sourceDirectory>
                         <sourceDirectory>scenarios/solrj-7.x-scenario</sourceDirectory>
-<!--                        <sourceDirectory>scenarios/spring-3.0.x-scenario</sourceDirectory>-->
+                        <sourceDirectory>scenarios/spring-3.0.x-scenario</sourceDirectory>
 <!--                        <sourceDirectory>scenarios/spring-3.1.x-scenario</sourceDirectory>-->
 <!--                        <sourceDirectory>scenarios/spring-4.1.x-scenario</sourceDirectory>-->
 <!--                        <sourceDirectory>scenarios/spring-4.3.x-scenario</sourceDirectory>-->
