+++ test/plugin/scenarios/retransform-class-scenario/src/main/java/org/apache/skywalking/apm/testcase/retransform/RetransformUtil.java
@@ -16,12 +16,12 @@
  *
  */
 
-package test.apache.skywalking.apm.testcase.retransform_class;
+package org.apache.skywalking.apm.testcase.retransform;
 
 import net.bytebuddy.agent.ByteBuddyAgent;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import test.apache.skywalking.apm.testcase.retransform_class.controller.CaseController;
+import org.apache.skywalking.apm.testcase.retransform.controller.CaseController;
 
 import java.lang.instrument.ClassFileTransformer;
 import java.lang.instrument.IllegalClassFormatException;
@@ -47,7 +47,7 @@ public class RetransformUtil {
                     int offset = indexOf(classfileBuffer, bytes);
                     if (offset != -1) {
                         byte[] replacingBytes = RETRANSFORM_VALUE.getBytes();
-                        System.arraycopy(replacingBytes,0, classfileBuffer, offset, replacingBytes.length);
+                        System.arraycopy(replacingBytes, 0, classfileBuffer, offset, replacingBytes.length);
                     }
                     return classfileBuffer;
                 }
@@ -71,10 +71,10 @@ public class RetransformUtil {
     }
 
     private static int indexOf(byte[] outerArray, byte[] smallerArray) {
-        for(int i = 0; i < outerArray.length - smallerArray.length+1; ++i) {
+        for (int i = 0; i < outerArray.length - smallerArray.length + 1; ++i) {
             boolean found = true;
-            for(int j = 0; j < smallerArray.length; ++j) {
-                if (outerArray[i+j] != smallerArray[j]) {
+            for (int j = 0; j < smallerArray.length; ++j) {
+                if (outerArray[i + j] != smallerArray[j]) {
                     found = false;
                     break;
                 }
