+++ test/plugin/scenarios/retransform-class-scenario/src/main/java/org/apache/skywalking/apm/testcase/retransform/controller/CaseController.java
@@ -16,7 +16,7 @@
  *
  */
 
-package test.apache.skywalking.apm.testcase.retransform_class.controller;
+package org.apache.skywalking.apm.testcase.retransform.controller;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -26,7 +26,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.client.HttpStatusCodeException;
-import test.apache.skywalking.apm.testcase.retransform_class.RetransformUtil;
+import org.apache.skywalking.apm.testcase.retransform.RetransformUtil;
 
 @Controller
 @RequestMapping("/case")
@@ -53,7 +53,4 @@ public class CaseController {
         // your codes
         return SUCCESS;
     }
-
-
-
 }
