+++ test/plugin/scenarios/retransform-class-tomcat-scenario/src/main/java/org/apache/skywalking/apm/testcase/retransform/RetransformUtil.java
@@ -16,7 +16,7 @@
  *
  */
 
-package test.org.apache.skywalking.apm.testcase.retransform_class;
+package org.apache.skywalking.apm.testcase.retransform;
 
 import net.bytebuddy.agent.ByteBuddyAgent;
 
@@ -43,7 +43,7 @@ public class RetransformUtil {
                     int offset = indexOf(classfileBuffer, bytes);
                     if (offset != -1) {
                         byte[] replacingBytes = RETRANSFORM_VALUE.getBytes();
-                        System.arraycopy(replacingBytes,0, classfileBuffer, offset, replacingBytes.length);
+                        System.arraycopy(replacingBytes, 0, classfileBuffer, offset, replacingBytes.length);
                     }
                     return classfileBuffer;
                 }
@@ -55,9 +55,7 @@ public class RetransformUtil {
             instrumentation.addTransformer(transformer, true);
             try {
                 instrumentation.retransformClasses(CaseController.class);
-                System.out.println("retransform classes success");
             } catch (Throwable e) {
-                System.out.println("retransform classes failure");
                 e.printStackTrace(System.out);
             }
 
@@ -68,10 +66,10 @@ public class RetransformUtil {
     }
 
     private static int indexOf(byte[] outerArray, byte[] smallerArray) {
-        for(int i = 0; i < outerArray.length - smallerArray.length+1; ++i) {
+        for (int i = 0; i < outerArray.length - smallerArray.length + 1; ++i) {
             boolean found = true;
-            for(int j = 0; j < smallerArray.length; ++j) {
-                if (outerArray[i+j] != smallerArray[j]) {
+            for (int j = 0; j < smallerArray.length; ++j) {
+                if (outerArray[i + j] != smallerArray[j]) {
                     found = false;
                     break;
                 }
